+++ src/IDBCursor.js
@@ -380,7 +380,7 @@ IDBCursor.prototype.__continueFinish = function (key, primaryKey, advanceState) 
     const me = this;
     const recordsToPreloadOnContinue = me.__advanceCount || CFG.cursorPreloadPackSize || 100;
     me.__gotValue = false;
-    me.__req.__readyState = 'pending'; // Unset done flag
+    me.__req.__done = false;
 
     me.__store.transaction.__pushToQueue(me.__req, function cursorContinue (tx, args, success, error, executeNextRequest) {
         function triggerSuccess (k, val, primKey) {