+++ src/IDBTransaction.js
@@ -86,10 +86,10 @@ IDBTransaction.prototype.__executeRequests = function () {
                 if (req) {
                     q.req = req; // Need to do this in case of cursors
                 }
-                if (q.req.__readyState === 'done') { // Avoid continuing with aborted requests
+                if (q.req.__done) { // Avoid continuing with aborted requests
                     return;
                 }
-                q.req.__readyState = 'done';
+                q.req.__done = true;
                 q.req.__result = result;
                 q.req.__error = null;
 
@@ -110,7 +110,7 @@ IDBTransaction.prototype.__executeRequests = function () {
                     // We've already called "onerror", "onabort", or thrown within the transaction, so don't do it again.
                     return;
                 }
-                if (q.req && q.req.__readyState === 'done') { // Avoid continuing with aborted requests
+                if (q.req && q.req.__done) { // Avoid continuing with aborted requests
                     return;
                 }
                 const err = findError(args);
@@ -119,7 +119,7 @@ IDBTransaction.prototype.__executeRequests = function () {
                     return;
                 }
                 // Fire an error event for the current IDBRequest
-                q.req.__readyState = 'done';
+                q.req.__done = true;
                 q.req.__error = err;
                 q.req.__result = undefined; // Must be undefined if an error per `result` getter
                 q.req.addLateEventListener('error', function (e) {
@@ -161,7 +161,7 @@ IDBTransaction.prototype.__executeRequests = function () {
                             q.op(tx, q.args, executeNextRequest, error);
                             return;
                         }
-                        if (q.req.__readyState === 'done') { // Avoid continuing with aborted requests
+                        if (q.req.__done) { // Avoid continuing with aborted requests
                             return;
                         }
                         q.op(tx, q.args, success, error, executeNextRequest);
@@ -409,7 +409,7 @@ IDBTransaction.prototype.__abortTransaction = function (err) {
 
         me.dispatchEvent(createEvent('__preabort'));
         me.__requests.filter(function (q, i, arr) { // eslint-disable-line promise/no-promise-in-callback
-            return q.req && q.req.__readyState !== 'done' && [i, -1].includes(
+            return q.req && !q.req.__done && [i, -1].includes(
                 arr.map((q) => q.req).lastIndexOf(q.req)
             );
         }).reduce(function (promises, q) {
@@ -418,7 +418,7 @@ IDBTransaction.prototype.__abortTransaction = function (err) {
             //  behaves first-in-first-out with the same timeout so we could
             //  just use a `forEach`.
             return promises.then(function () {
-                q.req.__readyState = 'done';
+                q.req.__done = true;
                 q.req.__result = undefined;
                 q.req.__error = createDOMException('AbortError', 'A request was aborted (an unfinished request).');
                 const reqEvt = createEvent('error', q.req.__error, {bubbles: true, cancelable: true});