+++ src/util.js
@@ -244,22 +244,27 @@ function defineListenerProperties (obj, listeners) {
     });
 }
 
-function defineReadonlyProperties (obj, props) {
+function defineReadonlyProperties (obj, props, getter = null) {
     props = typeof props === 'string' ? [props] : props;
     props.forEach(function (prop) {
-        Object.defineProperty(obj, '__' + prop, {
-            enumerable: false,
-            configurable: false,
-            writable: true
-        });
+        let o;
+        if (getter && prop in getter) {
+            o = getter[prop];
+        } else {
+            Object.defineProperty(obj, '__' + prop, {
+                enumerable: false,
+                configurable: false,
+                writable: true
+            });
+            // We must resort to this to get "get <name>" as
+            //   the function `name` for proper IDL
+            o = {
+                get [prop] () {
+                    return this['__' + prop];
+                }
+            };
+        }
 
-        // We must resort to this to get "get <name>" as
-        //   the function `name` for proper IDL
-        const o = {
-            get [prop] () {
-                return this['__' + prop];
-            }
-        };
         const desc = Object.getOwnPropertyDescriptor(o, prop);
         // desc.enumerable = true; // Default
         // desc.configurable = true; // Default